Rename images.ts
